Limit results to past 5 months
Now that we have a few reporting cycles of data it is taking a very long time to access this tool. Since the most we will ever need is 4 months plus a little extra while they go in and adjust data while preparing reports 5 months back should be a good working range.